звыполнено ДЗ модуляция
